refactor: Profile tab css 상 어색한 부분 수정
