+++ peloton-frontend/component/login/LoginNavigationRoot.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
-import ChangeProfile from "./login/ChangeProfile";
-import Login from "./login/Login";
-import ApplicationNavigationRoot from "./ApplicationNavigationRoot";
-import ErrorPage from "./ErrorPage";
+import ChangeProfile from "./ChangeProfile";
+import Login from "./Login";
+import ApplicationNavigationRoot from "../ApplicationNavigationRoot";
+import ErrorPage from "../ErrorPage";
 
 const LoginStack = createStackNavigator();
