--- peloton-frontend/component/profile/ProfileDefaultImage.js
@@ -1,22 +1,0 @@
-import { Image, StyleSheet } from "react-native";
-import React from "react";
-
-const ProfileDefaultImage = () => {
-  return (
-    <Image
-      style={styles.profileImage}
-      source={require("../../assets/default-profile.jpg")}
-    />
-  );
-};
-
-const styles = StyleSheet.create({
-  profileImage: {
-    resizeMode: "cover",
-    width: 100,
-    height: 100,
-    borderRadius: 100 / 2,
-  },
-});
-
-export default ProfileDefaultImage;
