+++ peloton-frontend/component/setting/SettingNavigationRoot.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
-import ProfileDetail from "./profile/profiledetail/ProfileDetail";
-import ProfileEdit from "./profile/profileedit/ProfileEdit";
-import CashUpdate from "./profile/profileedit/CashUpdate";
-import NameUpdate from "./profile/profileedit/NameUpdate";
-import License from "./setting/license/License";
-import Setting from "./setting/Setting";
+import ProfileDetail from "../profile/profiledetail/ProfileDetail";
+import ProfileEdit from "../profile/profileedit/ProfileEdit";
+import CashUpdate from "../profile/profileedit/CashUpdate";
+import NameUpdate from "../profile/profileedit/NameUpdate";
+import License from "./license/License";
+import Setting from "./Setting";
 
 const SettingStack = createStackNavigator();
